+++ src/main/java/br/com/itau/geradornotafiscal/model/ItemNotaFiscal.java
@@ -3,6 +3,8 @@ package br.com.itau.geradornotafiscal.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -16,12 +18,12 @@ public class ItemNotaFiscal {
     private String descricao;
 
     @JsonProperty("valor_unitario")
-    private double valorUnitario;
+    private BigDecimal valorUnitario;
 
     @JsonProperty("quantidade")
     private int quantidade;
 
     @JsonProperty("valor_tributo_item")
-    private double valorTributoItem;
+    private BigDecimal valorTributoItem;
 
 }
