+++ src/main/java/br/com/itau/geradornotafiscal/model/NotaFiscal.java
@@ -1,5 +1,6 @@
 package br.com.itau.geradornotafiscal.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -19,10 +20,10 @@ public class NotaFiscal {
     private LocalDateTime data;
 
     @JsonProperty("valor_total_itens")
-    private double valorTotalItens;
+    private BigDecimal valorTotalItens;
 
     @JsonProperty("valor_frete")
-    private double valorFrete;
+    private BigDecimal valorFrete;
 
     @JsonProperty("itens")
     private List<ItemNotaFiscal> itens;
