+++ src/main/java/br/com/itau/geradornotafiscal/service/NotaFiscalFactory.java
@@ -1,11 +1,11 @@
 package br.com.itau.geradornotafiscal.service;
 
-import br.com.itau.geradornotafiscal.model.Destinatario;
 import br.com.itau.geradornotafiscal.model.ItemNotaFiscal;
 import br.com.itau.geradornotafiscal.model.NotaFiscal;
 import br.com.itau.geradornotafiscal.model.Pedido;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -13,7 +13,7 @@ import java.util.UUID;
 @Component
 public class NotaFiscalFactory {
 
-    public NotaFiscal criar(Pedido pedido, List<ItemNotaFiscal> itensComTributo, double freteCalculado) {
+    public NotaFiscal criar(Pedido pedido, List<ItemNotaFiscal> itensComTributo, BigDecimal freteCalculado) {
         String idNotaFiscal = UUID.randomUUID().toString();
 
         return NotaFiscal.builder()
